+++ src/argilla/tasks/datasets/enums.py
@@ -12,19 +12,9 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-import typer
-
-from argilla.tasks.callback import init_callback
-from argilla.tasks.datasets.list import list_datasets
-from argilla.tasks.datasets.push import push_to_huggingface
-
-app = typer.Typer(
-    help="Holds CLI commands for datasets management", invoke_without_command=True, callback=init_callback
-)
-
-app.command(name="list", help="List datasets linked to user's workspaces")(list_datasets)
-app.command(name="push-to-huggingface", help="Push a dataset to HuggingFace Hub")(push_to_huggingface)
+from enum import Enum
 
 
-if __name__ == "__main__":
-    app()
+class DatasetType(str, Enum):
+    feedback = "feedback"
+    other = "other"